Moved JsonEqual to its own module.
